implement(AI_GoToFP): First implementation. Next: Testing.
