+++ Assets/GothicVR/Scripts/Npc/Actions/AnimationActions/GoToFp.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: be148037e77846af89eb7a43bfc912cf
+timeCreated: 1705250624
